Finished Namesspace changes and updated tests
