+++ src/Ardalis.Extensions/StringManipulation/Truncate.cs
@@ -1,0 +1,32 @@
+﻿using System;
+using System.Text.RegularExpressions;
+
+namespace Ardalis.Extensions.StringManipulation
+{
+    public static partial class StringManipulationExtensions
+    {
+        /// <summary>
+        /// Truncates string so that it is no longer than the specified number of characters. 
+        /// 
+        /// Source: https://stackoverflow.com/a/17249604/13729
+        /// </summary>
+        /// <param name="input">String to truncate.</param>
+        /// <param name="length">Maximum string length.</param>
+        /// <returns>Original string or a truncated one if the original was too long.</returns>
+        public static string Truncate(this string input, int length)
+        {
+            if (string.IsNullOrEmpty(input))
+            {
+                return string.Empty;
+            }
+
+            if (length < 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(length), "Length must be >= 0");
+            }
+
+            int maxLength = Math.Min(input.Length, length);
+            return input.Substring(0, maxLength);
+        }
+    }
+}
